backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.24.0

# ── Database ────────────────────────────────────────────────────────────────────
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# ── Auth ────────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# ── Config ──────────────────────────────────────────────────────────────────────
python-dotenv==1.0.0

# ── Cache ───────────────────────────────────────────────────────────────────────
redis==5.0.1

# ── HTTP client ─────────────────────────────────────────────────────────────────
httpx>=0.25.2

# ── Web Search ──────────────────────────────────────────────────────────────────
ddgs
duckduckgo-search
beautifulsoup4
lxml

# ── AI providers ────────────────────────────────────────────────────────────────
openai>=1.3.7
anthropic>=0.7.7
google-genai>=1.67.0

# ── Async ───────────────────────────────────────────────────────────────────────
# anyio version is managed by google-genai / httpx — do not pin manually
websockets>=13.0

# ── Document processing ─────────────────────────────────────────────────────────
pypdf==3.17.0
python-docx==1.1.2
aiofiles==23.2.1
Pillow>=10.0.0

# ── Validation ──────────────────────────────────────────────────────────────────
pydantic>=2.9.0
pydantic-settings>=2.4.0
email-validator==2.3.0

# ── Dev tools ───────────────────────────────────────────────────────────────────
pytest==7.4.3
black==23.11.0
